requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pyvis==0.3.2
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
networkx==3.2